Update push-a-sales-order-from-crm-to-desk.js
This script pushes a Sales Order from Zoho CRM to a new Sales Order inside Zoho Books. First, the script will check if the Account associated with the CRM Sales Order is a parent account or child account - if the account is a child account, the Sales Order will be pushed to books and billed to the parent account.